Refactored db entity.
